requirements.txt
# Core ML frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# API & Web
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Task Queue
celery>=5.3.0
redis>=5.0.0

# Model optimization
onnx>=1.14.0
onnxruntime>=1.16.0

# Utilities
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.66.0
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
httpx>=0.25.0
